+++ s4e-backend/src/test/java/pl/cyfronet/s4e/controller/ConfigControllerTest.java
@@ -2,6 +2,7 @@ package pl.cyfronet.s4e.controller;
 
 import org.junit.jupiter.api.Test;
 import org.springframework.beans.factory.annotation.Autowired;
+import org.springframework.beans.factory.annotation.Value;
 import org.springframework.boot.test.autoconfigure.web.servlet.AutoConfigureMockMvc;
 import org.springframework.test.web.servlet.MockMvc;
 import pl.cyfronet.s4e.BasicTest;
@@ -18,6 +19,9 @@ public class ConfigControllerTest {
     @Autowired
     private GeoServerProperties geoServerProperties;
 
+    @Value("${recaptcha.validation.siteKey}")
+    private String recaptchaSiteKey;
+
     @Autowired
     private MockMvc mockMvc;
 
@@ -25,6 +29,7 @@ public class ConfigControllerTest {
     public void shouldReturnConfiguration() throws Exception {
         mockMvc.perform(get(API_PREFIX_V1 + "/config"))
                 .andExpect(status().isOk())
-                .andExpect(jsonPath("$.geoserverUrl").value(geoServerProperties.getOutsideBaseUrl()));
+                .andExpect(jsonPath("$.geoserverUrl").value(geoServerProperties.getOutsideBaseUrl()))
+                .andExpect(jsonPath("$.recaptchaSiteKey").value(recaptchaSiteKey));
     }
 }