+++ CHANGELOG.md
@@ -26,6 +26,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 ### Fixed
 
 - Constraints names on Product and Scene tables [#364](https://github.com/cyfronet-fid/sat4envi/pull/364)
+- Build in private scope
 
 ### Updates