+++ CHANGELOG.md
@@ -25,6 +25,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ### Fixed
 
+- Return site-key instead of secret-key in ConfigController [#381](https://github.com/cyfronet-fid/sat4envi/issues/381)
 - Build in private scope [#372](https://github.com/cyfronet-fid/sat4envi/pull/372)
 - Constraints names on Product and Scene tables [#364](https://github.com/cyfronet-fid/sat4envi/pull/364)