+++ CHANGELOG.md
@@ -7,6 +7,8 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ## [Unreleased]
 
+## [v4.0.0]
+
 ### Added
 
 - Thumbnails MediaType and dimensions validation [#286](https://github.com/cyfronet-fid/sat4envi/issues/286)
@@ -35,4 +37,5 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 - 2020.03 backend dependencies update
 
-[unreleased]: https://github.com/cyfronet-fid/sat4envi/compare/0ebb1138...HEAD
+[unreleased]: https://github.com/cyfronet-fid/sat4envi/compare/v4.0.0...HEAD
+[v4.0.0]: https://github.com/cyfronet-fid/sat4envi/compare/0ebb1138...v4.0.0